# Openapi.yaml
...
openapi: 3.0.0
...
info:
  title: Api's test for login and registration-form and also convert json file to csv 
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 1.0.0
  contact:
    email: xyz.devoloper.com
  liscense:
    name: Apache 2.0
    url:  "http://localhost:8081"
servers:
  - url: http://localhost:8081
# body schema that referce the register form
...
paths:
...
 /api/jsonplaceHolder:
    get:
     summary: set the pagination 
     description: set the query as a page and limit
     parameters:
          - in: query
            name: page
            required: true
            schema:
              type: integer
              required: true
              example: 1   
          - in: query
            name: limit
            required: true
            schema:
              type: integer
              required: true
              example: 1   
          - in: query
            name: sortBy
            required: true
            schema:
              type: string
              required: true
              example: id_name   
          - in: query
            name: orderBy
            required: true
            schema:
              type: string
              required: true
              example: desc/acs   
 
     responses:
        200: 
          description: success
          content:
           application/json:
               schema:
                 type: object
                 properties:
                  previouse:
                     type: object
                     properties:
                       page:
                         type: interger
                         example: 2
                       limit: 
                         type: interger
                         example: 2
                  next:
                     type: object
                     properties:
                       page:
                         type: interger
                         example: 2
                       limit: 
                         type: interger
                         example: 2
                  data:
                     type: array
                     items:
                       page:
                         type: interger
                         example: 2
                       limit: 
                         type: interger
                         example: 2   
# pagination swapi
 /api/swapi:
     get:
      summary: set the pagination 
      description: set the query as a page and limit
      parameters:
           - in: query
             name: page
             required: true
             schema:
               type: integer
               required: true
               example: 1   
           - in: query
             name: limit
             required: true
             schema:
               type: integer
               required: true
               example: 1   
           - in: query
             name: sortBy
             schema:
               type: string
               example: id-name-title   
           - in: query
             name: orderBy
             schema:
               type: string
               example: asc/desc   
  
      responses:
         200: 
           description: success
           content:
            application/json:
                schema:
                  type: object
                  properties:
                   previouse:
                      type: object
                      properties:
                        page:
                          type: interger
                          example: 2
                        limit: 
                          type: interger
                          example: 2
                   next:
                      type: object
                      properties:
                        page:
                          type: interger
                          example: 2
                        limit: 
                          type: interger
                          example: 2
                   data:
                      type: array
                      items:
                        page:
                          type: interger
                          example: 2
                        limit: 
                          type: interger
                          example: 2                       